Decouple Brave Ads token redemption utils
